Divide test case
